+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Latest (Master)
+  * Added data import failure detection for when it's not possible to deduce the CSV delimiter
+
 # 0.2.1-rc2
   * Fixed issue causing propagation of importer properties to fail
   * Fixed issue with hiding the advanced/custom code/preview tabs in customizer